test_: add api client
